requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0

# AI and ML
langchain==0.0.335
openai==1.3.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
clip @ git+https://github.com/openai/CLIP.git
sentence-transformers==2.2.2

# Vector Database
pinecone-client==3.0.2

# AWS
boto3==1.29.3
botocore==1.32.3

# Utils
pillow==10.1.0
numpy==1.26.2
pandas==2.1.3
python-jose==3.3.0
httpx==0.25.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6 